pyproject.toml
[tool.poetry]
name = "JellyPlayer"
version = "1.0.0"
description = "Modern Jellyfin Client made with Python and Vue"
authors = ["Prayag <prayagprajapati.17@gmail.com>"]
license = "GPL-3.0-only"
readme = "README.md"
repository = "git@github.com:prayag17/JellyPlayer.git"
keywords = [
    "pyside6",
    "python",
    "jellyfin",
    "vue"
]

[tool.poetry.dependencies]
python = "~3.8"
PySide6 = "^6.3.0"
Flask = "^2.1.2"
waitress = "^2.1.1"
python-mpv = "^1.0.1"
pyinstaller = "^5.0.1"
PyOpenGL = "^3.1.6"

[tool.poetry.dev-dependencies]
pytest = "~5.2"
poethepoet = "^0.13.1"
autopep8 = "^1.6.0"

[tool.poe.tasks]
jellyplayer = "python jellyplayer/__init__.py"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"